gbuild: refactor gb_Library_set_componentfile:
Reverse the dependency: now the component target depends (order-only) on the
library, which should ensure that gb_CppunitTest_add_component can set up
dependencies such that the unit test does not run concurrently with
a library loaded via UNO being overwritten.

Unfortunately this requires introducing another phony dummy library
target, on which the module can depend (because we don't want to register
the component target as a module target).