prep work to eliminated Env.*.sh, introduce config_{host,build}.mk
This introduce config_host.mk and config_build.mk
generated from config_host.mk.in
The intent is for that file to be both sourcable in a shell and includable
in a gnu Makefile.
Eventually all the environment set-up will be in config_host.mk.in
and so there won't be a need to source anything in shell before make.