AutoText: fix segfault when delete/rename action applied while group selected.
In the Edit -> AutoText dialog, if an entry is selected and then a group
selected the delete/rename/etc actions are still shown in the "AutoText" menu
when they should not be. These actions are only applicable to entries and if
a user attempts to apply these actions while a group is selected the list box
becomes corrupted and leads to a segfault (the code for these actions
assumes the selected entry has a parent in the listbox).