fdo#36851, bnc#737190: Make the data validation popup more reliable.
Previously, the data validation message popup was not sticking on
reliably on Linux, and flickers very bad on Windows.  That was due to
the painting conflict between the top-most frame window and the grid
window.  By making the message popup window a child window of the
grid window, it paints more reliably.