ignore the (unsupported ) group customshape when exporting xlsx
currently the xml export code in oox doesn't handle grouped shapes very well, it produces an extra twoCellAnchor entry for the group ( which is incorrect ). This causes mso2010 to reject the content as corrupt. We need to handle grouping correctly ( e.g. a twoCellAnchor entry for the group and separate drawing shape (xdr:sp) elements for the shapes that are elements of the group ). As a stop gap we can just export the individual shapes of the group ( but not the grouping itself )