Attempt to manage without Berkeley DB on iOS and Android
Berkeley DB is used for help index and extension database. (Possibly
only for a backward-compatible format of the latter, I am not sure.)
Neither use makes much sense on Android and iOS.

The existing help is for LO on desktop OSes anyway, help for LO-based
apps on iOS and Android will naturally be quite different.

On iOS there will definitely be no "extensions", and probably we don't
want to bother with such on Android either.