SvxColumnItem: fix buggy STL conversion:
In the copy ctor, the vector must not be initialized with the constructor
that takes an integer, because that causes it to have that many
elements, which together with the loop that copies all source elements
leads to a vector with twice as many elements.
This causes a crash in writer when dragging the row separator line in a
table, because that code doesn't expect the item to have more elements
than the table has rows.
(regression from 0a82961765754bdc22932ba5fdc66f1bf2c5ba1f)