dmapper: Switch paragraphs to RTL based on the value of w:BiDi.
This is a fix for 43398: FORMATTING: Documents opened in LibreOffice Writer incorrectly appear as right justified.

Cause: dmapper used to treat the presence of a w:BiDi element in a paragraph's properties as an indicator of RTL directionality, which is incorrect. A w:BiDi  element may have a value of 0, indicating LTR.

Fix: Inspect the integral value of w:BiDi elements before switching to RTL.