n#693238, n#747471: create TableManager for inside shapes
When a texbox is contained inside a table cell... we were running into
some troubles when creating the table: the reference to the paragraphs
inside the textbox were passed to the TableManager. This was leading to
missing references when converting to a table because the paragraph
references within the textbox were dropped.