fdo#45962: fix context menu key yanking cursor out of header/footer
For the "context menu" key VCL synthesizes a cursor position that is the
center of the window, hence likely to be in the document body.
When SwEditWin::Command calls lcl_CheckHeaderFooterClick, it then moves
the document view cursor; prevent that by only calling that for actual
mouse events.  (regression from 9e756d80a80973ea88b4b4ab18580f7bbfcba105)