TextEngine::SeekCursor needs to call SetFont after all
31ed03016e627128033344a41ac65d9643903945 "callcatcher: build fixes" had removed
this with the comment "SetFont() doesn't do anything" but that is clearly wrong:
Without this, e.g. bold text within text fields is not displayed as such (e.g.,
"File - New - Templates and Documnets - Templates - Presentation Backgrounds -
Black and White": the captions in the right hand pane ("Title:", "Date:", etc.)
should be bold).