tubes: move Accept() logic into conference
This makes sense, because the flow for Accept() and Offer() is basically
symmetrical, and then conference becomes solely responsible for the
handling of channels once they've been requested/given to us.

While we're at it we also fix both code paths to work correctly if the
state becomes Open before Accept/Offer returns (which is possible).