Allow OServiceManagerWrapper to wrap all service manager implementations
...not just OServiceManager.  (This is a prerequisite to replace the
defaultBootstrap_initialComponentContext() implementation with a more performant
component context/service manager combo.)

The only reason for the restriction to OServiceManager apparently was the
additional createContentEnumeration overload with an XComponentContext
parameter.  It /looks/ to me like this is not really necessary, that it should
always be OK to instantiate implementation factories with this service manager
and its component context, instead of a component context (and its default
service manager) passed into a method call.  This would mean that the code could
be simplified further, needing a component context passed into a method only to
pass along to a new instance created via
createInstanceWith[ArgumentsAnd]Context, but then again I'm not 100% sure and
better leave it at that...

This also means that XUnoTunnel is gone from OServiceManager and XInitialization
is gone from OServiceManagerWrapper.