fdo#48969: switch units in Converter::convertDouble
The factor here is used to divide, so the parameters have to be
switched; this has always been broken but probably before
ad95aff7bcd2b68dcd105ac04a4a99ccdb6953a7 nothing called this.
This also reverts 488841dd64d384e94dda2b1a4a204e82e01b75e0 because
that was a workaround for the problem in the wrong place.