Resolves: fdo#39825 Make detection of certificate folder ui-configurable
Display in the (unix) options gui the NSS cert dir that has been auto-detected
for use with digital signatures. Show the other detected possibilities and
allow it to be overridden.

The autodetection should basically work out of the box, but if there's some
bustage at least the UI can be used to sort it out and/or let developers
debug it.

This removes the need for the horrible MOZILLA_CERTIFICATE_FOLDER hack,
though that's still supported.